+++ pyproject.toml
@@ -17,4 +17,4 @@ requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.scripts]
-sattern = "sattern.sattern:main"
+sattern = "sattern.src.sattern:main"
