+++ sattern/src/data/AAPL_1mo_history_data.json
