+++ sattern/src/data/AAPL_1y_history_data.json
