+++ sattern/src/data/NVDA_2y_history_data.json
